Added first version of text
After trying numerous glut solutions, the text doesn't want to cooperate
with legacy&modern GL. What I'm going to do is what I thought in the
first place: create a shape above the spheres and then texturize a side.
The problem with that, will be trying to orient the cube to always stare
at the player.